# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelMate API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface GeocodedLocation {
  city: string;
  country?: string;
  lat: number;
  lon: number;
  displayName: string;
}

export type PoiCategory = (typeof PoiCategory)[keyof typeof PoiCategory];

export const PoiCategory = {
  history: "history",
  gastronomy: "gastronomy",
  culture: "culture",
  nature: "nature",
  architecture: "architecture",
  nightlife: "nightlife",
  sports: "sports",
  relaxation: "relaxation",
  photography: "photography",
  landmark: "landmark",
  museum: "museum",
  park: "park",
  viewpoint: "viewpoint",
  religious: "religious",
  entertainment: "entertainment",
} as const;

export interface Poi {
  id: string;
  name: string;
  description?: string;
  category: PoiCategory;
  lat: number;
  lon: number;
  /** Duration in minutes */
  estimatedDuration: number;
  /** Cost in EUR */
  estimatedCost: number;
  isMustSee: boolean;
  isFree: boolean;
  optionalPaidExperience?: string | null;
  optionalPaidCost?: number | null;
  /** Score from 0-100 */
  popularityScore: number;
  imageUrl?: string | null;
  address?: string | null;
  openingHours?: string | null;
}

export type ItineraryRequestBudgetLevel =
  (typeof ItineraryRequestBudgetLevel)[keyof typeof ItineraryRequestBudgetLevel];

export const ItineraryRequestBudgetLevel = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type ItineraryRequestTravelRhythm =
  (typeof ItineraryRequestTravelRhythm)[keyof typeof ItineraryRequestTravelRhythm];

export const ItineraryRequestTravelRhythm = {
  relaxed: "relaxed",
  balanced: "balanced",
  dynamic: "dynamic",
} as const;

export type ItineraryRequestTravelProfile =
  (typeof ItineraryRequestTravelProfile)[keyof typeof ItineraryRequestTravelProfile];

export const ItineraryRequestTravelProfile = {
  solo: "solo",
  couple: "couple",
  family: "family",
  group: "group",
} as const;

export type ItineraryRequestTransportMode =
  (typeof ItineraryRequestTransportMode)[keyof typeof ItineraryRequestTransportMode];

export const ItineraryRequestTransportMode = {
  walking: "walking",
  car: "car",
  public_transport: "public_transport",
} as const;

export type ItineraryRequestInterestsItem =
  (typeof ItineraryRequestInterestsItem)[keyof typeof ItineraryRequestInterestsItem];

export const ItineraryRequestInterestsItem = {
  history: "history",
  gastronomy: "gastronomy",
  culture: "culture",
  nature: "nature",
  architecture: "architecture",
  nightlife: "nightlife",
  sports: "sports",
  relaxation: "relaxation",
  photography: "photography",
} as const;

export interface ItineraryRequest {
  city: string;
  country?: string;
  /**
   * @minimum 1
   * @maximum 14
   */
  days: number;
  budgetLevel: ItineraryRequestBudgetLevel;
  travelRhythm: ItineraryRequestTravelRhythm;
  travelProfile: ItineraryRequestTravelProfile;
  transportMode: ItineraryRequestTransportMode;
  interests: ItineraryRequestInterestsItem[];
}

export interface ItineraryDay {
  dayNumber: number;
  theme: string;
  pois: Poi[];
  /** Total duration in minutes */
  totalDuration: number;
  /** Total cost in EUR */
  totalCost: number;
  /** Encoded polyline for the route */
  routePolyline?: string | null;
}

export interface ItineraryResponse {
  city: string;
  country?: string;
  lat: number;
  lon: number;
  days: ItineraryDay[];
  totalEstimatedCost: number;
  generatedAt: string;
}

export type GeocodeCityParams = {
  city: string;
  country?: string;
};

export type GetPoisParams = {
  lat: number;
  lon: number;
  radius?: number;
};

export type GetPois200 = {
  pois: Poi[];
};