# lib/api-zod/src/generated/types/geocodedLocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelMate API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GeocodedLocation {
  city: string;
  country?: string;
  lat: number;
  lon: number;
  displayName: string;
}