# lib/api-zod/src/generated/types/itineraryDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Poi } from "./poi";

export interface ItineraryDay {
  dayNumber: number;
  theme: string;
  pois: Poi[];
  /** Total duration in minutes */
  totalDuration: number;
  /** Total cost in EUR */
  totalCost: number;
  /** Encoded polyline for the route */
  routePolyline?: string | null;
}