# lib/api-zod/src/generated/types/itineraryRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryRequestBudgetLevel } from "./itineraryRequestBudgetLevel";
import type { ItineraryRequestInterestsItem } from "./itineraryRequestInterestsItem";
import type { ItineraryRequestTransportMode } from "./itineraryRequestTransportMode";
import type { ItineraryRequestTravelProfile } from "./itineraryRequestTravelProfile";
import type { ItineraryRequestTravelRhythm } from "./itineraryRequestTravelRhythm";

export interface ItineraryRequest {
  city: string;
  country?: string;
  /**
   * @minimum 1
   * @maximum 14
   */
  days: number;
  budgetLevel: ItineraryRequestBudgetLevel;
  travelRhythm: ItineraryRequestTravelRhythm;
  travelProfile: ItineraryRequestTravelProfile;
  transportMode: ItineraryRequestTransportMode;
  interests: ItineraryRequestInterestsItem[];
}