# lib/api-zod/src/generated/types/itineraryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ItineraryDay } from "./itineraryDay";

export interface ItineraryResponse {
  city: string;
  country?: string;
  lat: number;
  lon: number;
  days: ItineraryDay[];
  totalEstimatedCost: number;
  generatedAt: string;
}