# lib/api-zod/src/generated/types/poi.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TravelMate API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PoiCategory } from "./poiCategory";

export interface Poi {
  id: string;
  name: string;
  description?: string;
  category: PoiCategory;
  lat: number;
  lon: number;
  /** Duration in minutes */
  estimatedDuration: number;
  /** Cost in EUR */
  estimatedCost: number;
  isMustSee: boolean;
  isFree: boolean;
  optionalPaidExperience?: string | null;
  optionalPaidCost?: number | null;
  /** Score from 0-100 */
  popularityScore: number;
  imageUrl?: string | null;
  address?: string | null;
  openingHours?: string | null;
}